requirements.txt
aiohappyeyeballs==2.4.1
aiohttp==3.10.6
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
et-xmlfile==1.1.0
frozenlist==1.4.1
fuzzywuzzy==0.18.0
idna==3.10
Levenshtein==0.26.0
lyricsgenius==3.0.1
multidict==6.1.0
numpy==2.1.1
openai==0.28.0
openpyxl==3.1.5
pandas==2.2.3
pandas-stubs==2.2.2.240909
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.0
pytz==2024.2
RapidFuzz==3.10.0
redis==5.0.8
requests==2.32.3
six==1.16.0
soupsieve==2.6
spotipy==2.24.0
tqdm==4.66.5
types-pytz==2024.2.0.20240913
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.12.1
